1. Support of VC2013 runtime. $(PlatformToolset) is added to path variables so it's easy to switch between 2010 and 2013 runtime (leads to changes in the names of folders: "v100", "v120" rather then VC2010, VC2013 in SupportingLibraries). Some new warnings caused by recently deprecated APIs are removed. Since we are not using Unicode, VC2013 additionally requires MBCS binaries: https://msdn.microsoft.com/en-us/library/dn251007.aspx. 2. Updates to make it compile with the latest GDAL_SDK.

+++ src/Ogr/Ogr2RawData.cpp
@@ -18,7 +18,7 @@ bool Ogr2RawData::Layer2RawData(OGRLayer* layer, Extent* extents, OgrDynamicLoad
 	layer->SetSpatialFilterRect(extents->left, extents->bottom, extents->right, extents->top);
 
 	loader->LockProvider(true);
-	int numFeatures = layer->GetFeatureCount();
+	int numFeatures = static_cast<int>(layer->GetFeatureCount());
 	loader->LockProvider(false);
 
 	callback->FeatureCount = numFeatures;
@@ -364,7 +364,7 @@ void Ogr2RawData::FieldsToShapeRecord(OGRFeatureDefn* poFields, OGRFeature* poFe
 		VARIANT* var = new VARIANT;
 		VariantInit(var);
 		var->vt = VT_I4;
-		var->lVal = poFeature->GetFID();
+		var->lVal = static_cast<long>(poFeature->GetFID());
 		data->Row->values.push_back(var);
 	}
